refactor: factor out getAppInstanceResources listener
Factor out the getAppInstanceResources listener to its own file.

refs #104
